+++ src/ts/Parsers/CometCampaignParser.ts
@@ -42,7 +42,8 @@ export class CometCampaignParser extends CampaignParser {
             creativeId: undefined,
             seatId: undefined,
             meta: json.meta,
-            session: session
+            session: session,
+            mediaId: response.getMediaId()
         };
 
         if(json && json.mraidUrl) {