+++ src/ts/Parsers/ProgrammaticMraidUrlParser.ts
@@ -36,7 +36,8 @@ export class ProgrammaticMraidUrlParser extends CampaignParser {
             creativeId: response.getCreativeId() || undefined,
             seatId: response.getSeatId() || undefined,
             meta: jsonMraidUrl.meta,
-            session: session
+            session: session,
+            mediaId: response.getMediaId()
         };
 
         const parameters: IMRAIDCampaign = {