+++ src/ts/Parsers/ProgrammaticStaticInterstitialParser.ts
@@ -21,7 +21,8 @@ export class ProgrammaticStaticInterstitialParser extends CampaignParser {
             creativeId: response.getCreativeId() || undefined,
             seatId: response.getSeatId() || undefined,
             meta: undefined,
-            session: session
+            session: session,
+            mediaId: response.getMediaId()
         };
 
         const displayInterstitialParams: IDisplayInterstitialCampaign = {