+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Pending
 
+* Initial support for realtime auction
+* Fix AdMob accelerometer fraud signals
+
+# Tue Mar 20 06:58:53 UTC 2018
+
+* Easter theme A/B test in groups 16 & 17
+* Remove overlay A/B test from groups 16 & 17
+
+# Mon Mar 19 07:48:44 UTC 2018
+
+* Support for new A/B testing framework
+* Add sessionDepth parameter to ad requests
+
+# Fri Mar 16 07:04:01 UTC 2018
+
 * iOS screen size fix for analytics
 * Split operative event manager into multiple operative event managers
 * XPromo: parse meta field